re-load grid driver after possibly changing range limits
The reason for this change is that the EtherCAN driver loads the range
limits from the proettcion database when it is starting up. The beta
collision test will (uusally) refresh the limits, to the FPU driver
needs to be initialized again to re-load the correct
limits. Otherwise, a protection error can be triggered.