Replace 'correct' function call with function handle
The image analysis functions are handling the distortion
correction by importing a function and calling it,
along with correction parameters. This happens for
every single point and is therefore quite slow.

In this change, use a passed function handle, or
generate that handle, to have a closure with correction
parameters, and apply this closure to the image coordinates
in question to get a faster computation.

The function handle is created by the new function
get_corection_func(), which accepts the parameters
which were before passed to each call of correct.

If no distortion coefficients are available, use
the PLATESCALE scaling parameter as a fall-back.
Using the fall-back requires that the parameter
is passed and defined.