Pass closure for distortion correction in dat_rep measurement
Generate and pass a closure as distortion correction function
in the datum repeatability measurement.

This requires that the required parameters are available
in the image analysis configuration parameters.

Note that the function still needs to be retrieved for
each individual mesaurement, because different
measurements can happen between changes to the
cameram, so that the calibration changes.