fix data flow in derivation of measurement set
The last change here introduced a buf for the case that there is a
non-empty measurement set and it is defined by the config file. In
this case, take the measurement configuration from the config file.

(The last change accounted for the fact that if the set if
selected tasks has an empty intersection with the
measurement tasks, the measurement set should be set to empty.
The way it was changed overlooked the dependency of
the measurement set on the configuration file.)