Fix: fixes bug in illumination for metrology height
Fix illumination bug in metrology height measurement.
Thos was tracked down to that the self-test for
metrology height is successful. For the measurements
however, all images are black. Apparently, a fundamental
assumption in the lamp control context manager does
not hold: the lamp settings are not composable.
More specifically, if the silhouette light is turned
on, and then the turntable moved, with the ambient
light turned on before starting the movement,
and the ambiend light turned off after the movement,
the silhouette light gets turned off, and remains so.
In effect, all images were completely black.

The fix consists into pulling the illumination
context manager for the image capture into
the movement loop for the turntable.