add mean alpha value for beta calibration measurement
Add the mean alpha value for the beta calibration
measurement. The alpha value adds up to the
actual angle of the beta arm in respect to
the camera axis. This is not noticable in the
calibration procedure itself, but makes
a difference when computing the reference
positions for the positional verification step.