insert missing transform to Cartesian coordinates
The coordinates returned by the image analysis functions
are OpenCV 'point' coordinates or pixel coordinates.
They are always positive with the origin or (0,0)
point in the upper left corner. Increasing y values
go downwards. This produces wrong rotations and angle signs
when, for example, applying cartesian2polar.

Fix this by flipping the sign of the y coordinate.
(Negatives values for y are correct).